add method for loading slanted dataset

+++ core.py
@@ -67,7 +67,6 @@ def load_mnist_data():
 
     return (x_train, y_train), (x_test, y_test), input_shape
 
-
 def load_straight_dataset(load_cached=False):
     """
     Parse out tens and ones digits from each image,
@@ -125,6 +124,72 @@ def load_straight_dataset(load_cached=False):
                     y_valid.append(ones_digit)
 
         with open(conf.OW_ULT_CHARGE_SHEARED_VALID_DATASET_PKL, 'wb') as f:
+            pickle.dump((x_valid, y_valid), f)
+            print('Done serializing validation dataset.')
+
+    x_valid = np.array(x_valid)
+    y_valid = np.array(y_valid)
+    n, h, w, c = _shape(x_valid)
+    x_valid = x_valid.reshape(n, h, w, c)
+    y_valid = keras.utils.to_categorical(y_valid, conf.NUM_CLASSES)
+
+    return x_valid, y_valid
+
+
+def load_slanted_dataset(load_cached=False):
+    """
+    Parse out tens and ones digits from each image.
+    """
+    tens_bbox = Bbox(x=2, y=0, w=15, h=30)
+    ones_bbox = Bbox(x=17, y=0, w=15, h=30)
+    solo_ones_bbox = Bbox(x=0, y=0, w=23, h=30)
+    ult_charge_bbox = Bbox(x=625, y=590, w=30, h=30)
+    warped_size = (28, 28)
+
+    if load_cached and os.path.exists(conf.OW_ULT_CHARGE_SLANTED_VALID_DATASET_PKL):
+        with open(conf.OW_ULT_CHARGE_SLANTED_VALID_DATASET_PKL, 'rb') as f:
+            x_valid, y_valid = pickle.load(f)
+            print('Deserialized validation dataset from Pickle file.')
+    else:
+        x_valid = []
+        y_valid = []
+        directories = [file for file in os.listdir(conf.OW_ULT_CHARGE_EVAL_DATASET_DIR)\
+                       if os.path.isdir(os.path.join(conf.OW_ULT_CHARGE_EVAL_DATASET_DIR, file))]
+        ult_dirs = []
+        for file in directories:
+            m = re.search('(^\d{1,3}$)', file)
+            if m and m.groups():
+                ult_dirs.append(file)
+
+        for dirname in ult_dirs:
+            ult_charge = int(dirname)
+            tens_digit = ult_charge // 10
+            ones_digit = ult_charge % 10
+
+            print('Capturing {} and {} for {}'.format(tens_digit, ones_digit, dirname))
+            directory = os.path.join(conf.OW_ULT_CHARGE_EVAL_DATASET_DIR, dirname)
+
+            for file in os.listdir(directory):
+                img = mpimg.imread(os.path.join(directory, file))
+                region = crop_region(img, ult_charge_bbox)
+
+                if tens_digit == 0:
+                    digit = cv2.resize(crop_region(region, solo_ones_bbox), \
+                                            warped_size, interpolation=cv2.INTER_LINEAR)
+                    digit = color.rgb2gray(digit)
+                    x_valid.append(digit)
+                    y_valid.append(ones_digit)
+                else:
+                    tens = cv2.resize(crop_region(region, tens_bbox), warped_size, interpolation=cv2.INTER_LINEAR)
+                    ones = cv2.resize(crop_region(region, ones_bbox), warped_size, interpolation=cv2.INTER_LINEAR)
+                    tens = color.rgb2gray(tens)
+                    ones = color.rgb2gray(ones)
+                    x_valid.append(tens)
+                    y_valid.append(tens_digit)
+                    x_valid.append(ones)
+                    y_valid.append(ones_digit)
+
+        with open(conf.OW_ULT_CHARGE_SLANTED_VALID_DATASET_PKL, 'wb') as f:
             pickle.dump((x_valid, y_valid), f)
             print('Done serializing validation dataset.')
